requirements.txt
# Core ML
torch>=2.6.0
torchvision>=0.21.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0

# Explainability
shap>=0.44.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# AWS SageMaker
boto3>=1.34.0
sagemaker>=2.207.0

# MLflow experiment tracking
mlflow>=2.10.0

# Data validation
great-expectations>=0.18.0
pydantic>=2.6.0

# API serving
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.0
PyYAML>=6.0.1
joblib>=1.3.2

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
moto>=5.0.0